+++ src/Recyclarr.TrashGuide/CustomFormat/CustomFormatLoader.cs
@@ -9,11 +9,13 @@ public class CustomFormatLoader : ICustomFormatLoader
 {
     private readonly ServiceJsonLoader _loader;
     private readonly ICustomFormatCategoryParser _categoryParser;
+    private readonly IFileSystem _fs;
 
-    public CustomFormatLoader(ServiceJsonLoader loader, ICustomFormatCategoryParser categoryParser)
+    public CustomFormatLoader(ServiceJsonLoader loader, ICustomFormatCategoryParser categoryParser, IFileSystem fs)
     {
         _loader = loader;
         _categoryParser = categoryParser;
+        _fs = fs;
     }
 
     public ICollection<CustomFormatData> LoadAllCustomFormatsAtPaths(
@@ -23,8 +25,9 @@ public class CustomFormatLoader : ICustomFormatLoader
         var categories = _categoryParser.Parse(collectionOfCustomFormats).AsReadOnly();
         return _loader.LoadAllFilesAtPaths<CustomFormatData>(jsonPaths, x => x.Select(cf =>
         {
-            var matchingCategory = categories.FirstOrDefault(
-                y => y.CfName.EqualsIgnoreCase(cf.Obj.Name) || y.CfAnchor.EqualsIgnoreCase(cf.File.Name));
+            var matchingCategory = categories.FirstOrDefault(y =>
+                y.CfName.EqualsIgnoreCase(cf.Obj.Name) ||
+                y.CfAnchor.EqualsIgnoreCase(_fs.Path.GetFileNameWithoutExtension(cf.File.Name)));
 
             return cf.Obj with
             {